fix: add missing file list update after undo delete in space
(cherry picked from commit d416ac4474c7122aaa96d74c3ee1a4f64b427d3b)
