+++ packages/web-pkg/src/helpers/resource/index.ts
@@ -2,4 +2,5 @@ export * from './sameResource'
 export * from './conflictHandling'
 export * from './filter'
 export * from './icon'
+export * from './isItemInCurrentFolder'
 export * from './renameResource'
